refactor(AlertSummary): Break out nested modules
